# User login scoping (API)