~~**NOTE:**  Currently built on top of https://github.com/ManageIQ/manageiq/pull/20471~~

~~(though, this probably could be done without, but worth noting the **Benchmarks** included those changes)~~

A collection of tweaks to the `User` and `Authenticator::*` models to enable more effecient querying of `User` and related models

Requires `manageiq-api` changes to take effect:  https://github.com/ManageIQ/manageiq-api/pull/888

Benchmarks
----------

Note:  Improvements are in number of **queries**.


### Before

```
$ bundle exec miqperf benchmark -ac2 "/api/users"
$ bundle exec miqperf report --last
/api/auth
|   ms | queries | query (ms) | rows |
| ---: |    ---: |       ---: | ---: |
| 5088 |      26 |        183 |   10 |
/api/users
|   ms | queries | query (ms) | rows |
| ---: |    ---: |       ---: | ---: |
|  267 |      12 |         13 | 1494 |
|   17 |      10 |        3.9 |   12 |
```


### After

```
$ bundle exec miqperf benchmark -ac2 "/api/users"
$ bundle exec miqperf report --last
/api/auth
|   ms | queries | query (ms) | rows |
| ---: |    ---: |       ---: | ---: |
| 4949 |      12 |        181 |    1 |
/api/users
|   ms | queries | query (ms) | rows |
| ---: |    ---: |       ---: | ---: |
|  258 |       7 |         12 | 1489 |
|   15 |       5 |        3.3 |    7 |
```


Links
-----

* https://github.com/ManageIQ/manageiq-api/pull/888
* More improvements as part of this effort:  https://github.com/ManageIQ/manageiq-api/issues/880