@NickLaMuro 2020-10-12
> Apologies for the late reply...

@Fryguy no worries, though this is probably going to take a back seat since I don't really have a need for this at the moment, and I pushed it off into https://github.com/ManageIQ/manageiq-api/issues/889 as well.

Also... I really don't want to bother with rebasing it at the moment...