@NickLaMuro 2021-01-25
Okay, I wouldn't mind waiting until CI does pass, but I think this failure is from Libor's changes (and by extension, @kbrock ):

https://travis-ci.com/github/ManageIQ/manageiq/jobs/474546624#L791-L802

So I will wait to see if master starts passing before I try again.