@kbrock 2021-02-05
you going to address all the hashrockets - or you want to ship as is

ruby 3.0 does need colons and not hash rockets