@NickLaMuro 2021-02-05
> you going to address all the hashrockets - or you want to ship as is

All of these are for method arguments that are `kwargs`, which is required.  As Jason [said in gitter](https://gitter.im/ManageIQ/manageiq/core?at=601d8f1ac399ef60aff4d84e):

> I've always said that the `arg:` value syntax made sense for keyword args , and it should not have been reused for hashes 😉

and I agree.  Also, as you said (I guess...)

> ruby 3.0 does need colons and not hash rockets [for kwargs]

Which is the only place these `rubocop` warnings are from, so they are irrelevant.

This is also a style for `kwargs` that is used elsewhere in the application:

https://github.com/ManageIQ/manageiq/blob/72fa6ef5e0419ebbb680e1e5ce1f5d109c44d99b/lib/ansible/runner.rb#L15

So I am not doing anything non-standard here.