@Fryguy 2021-02-11
I wonder if rubocop has a way to differentiate Hashes and keyword args (particularly because of the upcoming 3.0 changes)